refactor(search): extract pure algorithm into search-core
Splits the prefix / rotated-token / substring search algorithm out of
search-index.ts (which depends on fs/zlib) into a Node-free search-core
module. Server-side search-index.ts re-exports the pure pieces and
keeps the disk loader; client-side code added in the next commit will
import directly from search-core.

Existing search-index tests still pass against the re-exported API.
